pyproject.toml
[project]
name = "jarvis"
version = "0.1.0"
description = "WhatsApp Claude Code Agent"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=2.31.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "pandas>=3.0.0",
    "pyarrow>=23.0.0",
    "python-crontab>=3.3.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]

[project.scripts]
jarvis = "jarvis.main:main"

[tool.setuptools.packages.find]
include = ["jarvis*"]
